Recreate controls on next set if gamma size is zero
If the output is currently disabled and has no CRTC, the exact gamma
details might not yet be known by the compositor, but compositors do not
send updates if the value changes.

Destroy the gamma control and let it be recreated on next temperature
set, hoping that we eventually get functioning size.