Check the result of realloc, exit if failing
If realloc fails and returns NULL, the char pointer allocated by strdup
would be written to `(vec->len - 1) * sizeof(char*)`. Depending on the
state of `vec->len` (equal to the number of `-o` arguments wlsunset was
instantiated with) at the time of failure, this will either lead to a
segfault or a memory corruption during wlsunset startup.

Exit with an error if allocation failed.